Refactor StudentLogin component
Refactor the StudentLogin component into smaller, more manageable files for improved code organization and maintainability.
