+++ src/pages/student/StudentLogin.tsx
@@ -1,129 +1,22 @@
+
 import React, { useState } from 'react';
-import { useNavigate, Link } from 'react-router-dom';
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Button } from '@/components/ui/button';
-import { Form, FormControl, FormDescription, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
-import { useToast } from '@/components/ui/use-toast';
-import { z } from 'zod';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useForm } from 'react-hook-form';
-import { supabase } from '@/integrations/supabase/client';
-import { GraduationCap, LogIn, UserPlus } from 'lucide-react';
+import { Link } from 'react-router-dom';
+import { Card, CardContent, CardFooter, CardHeader, CardTitle, CardDescription } from '@/components/ui/card';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-
-const loginSchema = z.object({
-  email: z.string().email({ message: "Please enter a valid email address" }),
-  password: z.string().min(6, { message: "Password must be at least 6 characters" }),
-});
-
-const registerSchema = z.object({
-  fullName: z.string().min(3, { message: "Full name must be at least 3 characters" }),
-  email: z.string().email({ message: "Please enter a valid email address" }),
-  password: z.string().min(6, { message: "Password must be at least 6 characters" }),
-  rollNumber: z.string().min(1, { message: "Roll number is required" }),
-  className: z.string().min(1, { message: "Class is required" }),
-});
-
-type LoginFormValues = z.infer<typeof loginSchema>;
-type RegisterFormValues = z.infer<typeof registerSchema>;
+import { GraduationCap } from 'lucide-react';
+import { StudentLoginForm } from '@/components/student/StudentLoginForm';
+import { StudentRegisterForm } from '@/components/student/StudentRegisterForm';
+import { useStudentAuth } from '@/hooks/useStudentAuth';
+import { LoginFormValues, RegisterFormValues } from '@/schemas/studentAuth';
 
 const StudentLogin = () => {
-  const [loading, setLoading] = useState(false);
-  const navigate = useNavigate();
-  const { toast } = useToast();
+  const { loading, handleLogin, handleRegister } = useStudentAuth();
+  const [registeredEmail, setRegisteredEmail] = useState("");
   
-  const loginForm = useForm<LoginFormValues>({
-    resolver: zodResolver(loginSchema),
-    defaultValues: {
-      email: "",
-      password: "",
-    },
-  });
-
-  const registerForm = useForm<RegisterFormValues>({
-    resolver: zodResolver(registerSchema),
-    defaultValues: {
-      fullName: "",
-      email: "",
-      password: "",
-      rollNumber: "",
-      className: "",
-    },
-  });
-
-  const handleLogin = async (data: LoginFormValues) => {
-    setLoading(true);
-    try {
-      const { error } = await supabase.auth.signInWithPassword({
-        email: data.email,
-        password: data.password,
-      });
-
-      if (error) throw error;
-      
-      toast({
-        title: "Login successful",
-        description: "Welcome back to your student dashboard",
-      });
-      
-      navigate('/student/dashboard');
-    } catch (error) {
-      toast({
-        variant: "destructive",
-        title: "Login failed",
-        description: error.message || "Please check your credentials and try again",
-      });
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleRegister = async (data: RegisterFormValues) => {
-    setLoading(true);
-    try {
-      // Register user with Supabase Auth
-      const { data: authData, error: authError } = await supabase.auth.signUp({
-        email: data.email,
-        password: data.password,
-        options: {
-          data: {
-            full_name: data.fullName,
-          }
-        }
-      });
-
-      if (authError) throw authError;
-
-      if (authData?.user) {
-        // Add student profile data to the students table
-        const { error: profileError } = await supabase
-          .from('students')
-          .insert({
-            user_id: authData.user.id,
-            full_name: data.fullName,
-            roll_number: data.rollNumber,
-            class_name: data.className,
-            email: data.email,
-          });
-
-        if (profileError) throw profileError;
-      }
-
-      toast({
-        title: "Registration successful",
-        description: "Your student account has been created. Please check your email for verification.",
-      });
-      
-      loginForm.setValue('email', data.email);
-    } catch (error) {
-      toast({
-        variant: "destructive",
-        title: "Registration failed",
-        description: error.message || "An error occurred during registration",
-      });
-    } finally {
-      setLoading(false);
+  const onRegisterSubmit = async (data: RegisterFormValues) => {
+    const email = await handleRegister(data);
+    if (email) {
+      setRegisteredEmail(email);
     }
   };
 
@@ -148,136 +41,32 @@ const StudentLogin = () => {
           
           <TabsContent value="login">
             <CardHeader>
-              <CardTitle className="flex items-center">
-                <LogIn className="mr-2 h-5 w-5" />
-                Student Login
-              </CardTitle>
+              <CardTitle>Student Login</CardTitle>
               <CardDescription>
                 Enter your credentials to access your student dashboard
               </CardDescription>
             </CardHeader>
             <CardContent>
-              <Form {...loginForm}>
-                <form onSubmit={loginForm.handleSubmit(handleLogin)} className="space-y-4">
-                  <FormField
-                    control={loginForm.control}
-                    name="email"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Email</FormLabel>
-                        <FormControl>
-                          <Input placeholder="student@example.com" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={loginForm.control}
-                    name="password"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Password</FormLabel>
-                        <FormControl>
-                          <Input type="password" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <Button type="submit" className="w-full" disabled={loading}>
-                    {loading ? "Logging in..." : "Login"}
-                  </Button>
-                </form>
-              </Form>
+              <StudentLoginForm 
+                onSubmit={handleLogin} 
+                loading={loading}
+                defaultEmail={registeredEmail}
+              />
             </CardContent>
           </TabsContent>
           
           <TabsContent value="register">
             <CardHeader>
-              <CardTitle className="flex items-center">
-                <UserPlus className="mr-2 h-5 w-5" />
-                Student Registration
-              </CardTitle>
+              <CardTitle>Student Registration</CardTitle>
               <CardDescription>
                 Create a new student account to access your academic information
               </CardDescription>
             </CardHeader>
             <CardContent>
-              <Form {...registerForm}>
-                <form onSubmit={registerForm.handleSubmit(handleRegister)} className="space-y-4">
-                  <FormField
-                    control={registerForm.control}
-                    name="fullName"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Full Name</FormLabel>
-                        <FormControl>
-                          <Input placeholder="John Doe" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={registerForm.control}
-                    name="email"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Email</FormLabel>
-                        <FormControl>
-                          <Input placeholder="student@example.com" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <FormField
-                    control={registerForm.control}
-                    name="password"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormLabel>Password</FormLabel>
-                        <FormControl>
-                          <Input type="password" {...field} />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                  <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-                    <FormField
-                      control={registerForm.control}
-                      name="rollNumber"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Roll Number</FormLabel>
-                          <FormControl>
-                            <Input placeholder="R-12345" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                    <FormField
-                      control={registerForm.control}
-                      name="className"
-                      render={({ field }) => (
-                        <FormItem>
-                          <FormLabel>Class</FormLabel>
-                          <FormControl>
-                            <Input placeholder="Grade 10" {...field} />
-                          </FormControl>
-                          <FormMessage />
-                        </FormItem>
-                      )}
-                    />
-                  </div>
-                  <Button type="submit" className="w-full" disabled={loading}>
-                    {loading ? "Registering..." : "Register"}
-                  </Button>
-                </form>
-              </Form>
+              <StudentRegisterForm 
+                onSubmit={onRegisterSubmit} 
+                loading={loading}
+              />
             </CardContent>
           </TabsContent>
         </Tabs>
