+++ src/schemas/studentAuth.ts
@@ -1,0 +1,18 @@
+
+import { z } from 'zod';
+
+export const loginSchema = z.object({
+  email: z.string().email({ message: "Please enter a valid email address" }),
+  password: z.string().min(6, { message: "Password must be at least 6 characters" }),
+});
+
+export const registerSchema = z.object({
+  fullName: z.string().min(3, { message: "Full name must be at least 3 characters" }),
+  email: z.string().email({ message: "Please enter a valid email address" }),
+  password: z.string().min(6, { message: "Password must be at least 6 characters" }),
+  rollNumber: z.string().min(1, { message: "Roll number is required" }),
+  className: z.string().min(1, { message: "Class is required" }),
+});
+
+export type LoginFormValues = z.infer<typeof loginSchema>;
+export type RegisterFormValues = z.infer<typeof registerSchema>;
